+++ test/test33.c
@@ -49,7 +49,7 @@ int main()
             prints("Error at offset ");
             printi(index);
             prints(" ");
-            printi(datap[index]);
+            printi((int)datap[index]);
             prints("\n");
         }
         else
